Moved util functions into their own function
